requirements.txt
streamlit
torch
transformers>=4.30.0
pillow
accelerate
datasets
evaluate
